requirements.txt
pandas
numpy
psycopg2-binary
pymongo
sqlalchemy
scikit-learn
pyspark
tqdm
faiss-cpu